Fix large literal integer cannot be used in IDL.
Because the range of haskell's Int data type is implementation specific, so it is possible to fail handle large integer such as 64bit integer values or 32bit unsigned integer values. By the way, the range in Haskell Platform 2012.2 on Windows is 32bit signed integer value.
This commit changes the underlying value type of LInt to Integer.
Note that language binding might not assume that LInt value can be 64bit integer, so they still do not work for 64bit integer literal because some language (Java, for example) require using special notation for 64bit integer literals like suffix.